Merged PR 785239: [Linux sandbox] Resolve intermediate symlinks for all report paths, and add related tests
Make us resolve all paths that we report back to the sandbox (except the explicit cases where we don't want to follow the last symlink, like readlinks or stats). Also makes our intermediate symlink resolution logic preserve errno to not mess with callers.

Additionally, make the SandboxEvent immutable after constructing a report, so we don't accidentally rely on changing something insde it that would be inconsistent with the report we then create

Also added some tests for it.

Related work items: #2179576