Merged PR 788728: Make sure that RetryInfo is set for cancelled remote pips
It appears that when we stop a worker due to an internal error, we mark all pips as cancelled. For pips, that have not started running, we were setting Cancelled status, but were not setting RetryInfo. When the orchestrator receives the result back, it's expecting to see RetryInfo unless the orchestrator is terminating.  In case of low disk space, we stop a worker, but not the orchestrator, so we end up with a pip that breaks the assert, and we crash.

Related work items: #2185688