Merged PR 788408: Refactor GitInfoManager and auto-collect git repo for /traceinfo:codebase
Refactored GitInfoManager to avoid the need of LoggingContext as an input.
We will be logging the related failures or gitRepoInfo later, after the creation of loggingContext.

This way we can set the gitRemoteRepoUrl as the codebase property in case its not passed by the user or obtained in ADO.

After:
Case 1: No value provided by the user

dominoinvocation
| where SessionId == "8b74c916-0000-0000-bd0d-409fefa41fc4"
| project GitRemoteRepoUrl, Environment, CommandLine, UserName

Case 2: The user provided value always overrides this value
passed /traceInfo:codebase=TestBXLGitInfo to the build.

dominoinvocation
| where SessionId == "02181b5e-0000-0000-bd0d-1eabceb9d073"
| project GitRemoteRepoUrl, Environment, CommandLine, UserName

Related work items: #2171047