Merged PR 788685: Modify last access time tracking so it's always updated'
This change modifies how we touch files when doing uploads of files that already exist or pins. Currently, we'll update the last access time by doing a 1-byte download.

In the new version, we instead update the Content-Language to a random value. The reason this works is that updating Content-Language updates the Last Modification Time, the ETag, and the Last Access Time. Because this only works when we have write permissions, we do this only for non-read-only sessions.

This also changes GC so that when it reads the last access time from a blob, it also fetches the ETag and the last modification time.

The reason this change is being made is we've had already 6 WDG build failures where:
1. A content is added long ago
2. A build runs, it inserts the content, the content is already there, so we touch it instead
3. In parallel, GC is running and sees the last access time from before the touch, so it proceeds to delete it. Because we had no way to prevent the deletion from happening when there's a race, this would cause a CTMIS

Reverts !785193

Reverts !788669

This re-introduces the change along with a fix to prevent build failures due to the ETag changing while a parallel download is happening in parallel and already downloaded one of the chunks.