Merged PR 788468: Don't save fingerprint salt as part of cached graph
The fingerprint salt specified on the command line wasn't always kicking it. It turns out this is because the salt is serialized with ConfigFileState and reused when the graph is reloaded.

I don't know why that salt was ever serialized. The functionality goes way back:
https://mseng.visualstudio.com/Domino/_git/Domino/commit/a33bd9223ed81c5737661bbc8ba568b95be67864?refName=refs/heads/master&path=/Public/Src/Engine/Domino.Engine/ConfigFileState.cs

Separately, there is some fingerprint salt serialization that happens as part of incremental scheduling. That prevented this bug from kicking in for local builds. That already correctly wires in the salt in its graph eviction logic so there is no issue there.

Related work items: #2184764