Merged PR 784812: Add a DebugTrace struct for collecting extra logs in investigations, and add some of this tracing to ObservedInputProcessor
The point of this abstraction is to to provide a way to add debugging capabilities throughout the code that will have minimal overhead when in a 'disabled' state. This way, consumers only need to worry about this "construction" time  (typically, using an optional setting to enable the debugging in that particular section of the code), and can add arbitrary debugging  lines in the rest of the logic without worrying about impacting builds that are not meant to be observed with the extra logging. This is achieved by using a custom InterpolatedStringHandler, which will only construct the interpolated strings when the debugging is enabled.

I ran a Benchmark.NET that does some random processing of a bunch of strings:
```cs
int containsA = 0, randomCondition = 0;
// myGuidStrings contains 5000 guids as strings
for (int i = 0; i < myGuidStrings.Count; i++)
{
        using var trace = new DebugTrace(enabled);
	trace.AppendLine("Starting processing");
	var s = myGuidStrings[i];
	var someRandomCondition = s.Contains("a477");
	trace.AppendLine($"Condition: {someRandomCondition}");
	if (someRandomCondition){
		randomCondition++;
	}
	var splits = s.Split("8");
	foreach (var ss in splits)
	{
		trace.AppendLine($"split {ss}");
		if (ss.Contains('a'))
		{
			trace.AppendLine("split contains a");
			containsA++;
		}
	}

	diag = trace.ToString();
}
```

The benchmark runs with

```
(1) Enabled:  Constructing DebugTrace with enabled = true
(2) Disabled: Constructing DebugTrace with enabled = false
(3) NoTrace:  DebugTrace stripped from the code (not using it at all)
```

With 5000 strings:
```
| Method   | Mean       | Error    | StdDev  | Gen0     | Allocated  |
|--------- |-----------:|---------:|--------:|---------:|-----------:|
| Enabled  | 2,976.2 us | 11.27 us | 9.99 us | 203.1250 | 3344.74 KB |
| Disabled |   888.7 us |  3.68 us | 3.26 us |  52.7344 |  868.15 KB |
| NoTrace  |   872.0 us |  3.05 us | 2.85 us |  52.7344 |  868.98 KB |
```

With 100,000 strings:

```
| Method   | Mean     | Error    | StdDev   | Gen0      | Allocated |
|--------- |---------:|---------:|---------:|----------:|----------:|
| Enabled  | 63.15 ms | 0.323 ms | 0.302 ms | 4000.0000 |  65.44 MB |
| Disabled | 19.30 ms | 0.224 ms | 0.209 ms | 1062.5000 |     17 MB |
| NoTrace  | 17.70 ms | 0.133 ms | 0.124 ms | 1062.5000 |  17.03 MB |
```

Comparing `Disabled` and `NoTrace` shows that having this code in the "disabled" state has practically no overhead (consider that for the benchmark, like ~half the operations are on the DebugTrace object), and no extra allocations.