Merged PR 789992: Do not throttle stdout/err materialization
We are seeing some 1JS builds that spend a quite amount of time in ReplayWarningsFromCacheAsync (some up to 20 minutes) due to waiting for stderr/out to materialization semaphore. Make those materialization requests so they don't need to acquire the materialization semaphore. These are always 2 materialization requests per pip (so in terms of throttling they should be negligible), but at the same time this step is blocking the scheduler.

A more comprehensive change should include regular input/output materialization, such that we have some sort of distinct queues/priority queues for those. But that's a bigger change to make.