Merged PR 789674: Fix RegexPathMatch for path prefixes in Windows
Adding back this PR with a small fix.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/784626

As per the current PR. We first check if there is a match with the Windows Prefix if not we remove the prefix and then match it again to see if there is a match.

The first test case in the newly added unit test validates the change
 [InlineData(@"^\\\\\?\\.*", @"\\?\c:\foo\bar\file.txt", true)]

Related work items: #2165593