Merged PR 791180: Refactor and fix capture build properties for org and codebase for Reliability dashboard
Refactor and fix capture build properties for org and codebase for Reliability dashboard.
Most of the URL's in Reliability dashboard are of the format - dev.azure.com and visualstudio.com
So I reused the existing code to normalize gitRemoteRepoUrl's/
Added few test cases to test these scenarios.

Related work items: #2188136