Merged PR 794662: Fix msbuild frontend when building solution
When building from solution, MSBuild frontend can easily create a path segment that exceeds the segment limit (> 256). This is because the frontend tries to create a unique segment by appending global properties. Even though long paths is enabled, the segment limit cannot be changed.

In this PR, we simply hash the path segment. Due to hashing, the user may need to look at the log file to correlate the path and the pip.

This PR also addresses some issue with path ending with backslash appearing in cmd line. Apparently there's a bug in MSBuild that includes the ending quote as part of the path.

Related work items: #2191024