Merged PR 793701: Speed up ObjectPool
ObjectPool is backed by a ConcurrentStack. When the ObjectPool returns an item to the pool, it checks the stack length to see if it should hold onto it or not. The implementation of ConcurrentStack.Count is o(n). This change adds parallel tracking with a counter to keep from iterating that stack over and over. This also removes some of the Interlocked calls used for tracking statistics. It is not important that those are always synchronized.

Accessing a pool 1 million times on a 16 core dev machine:
LoopParallelism - number of threads concurrently accessing the pool
PoolSize - number of items in the ObjecPool

| Method   | LoopParallelism | PoolSize | Mean      | Error    | StdDev    | Median    |
|--------- |---------------- |--------- |----------:|---------:|----------:|----------:|
| Original | 1               | 64       |  30.61 ms | 0.182 ms |  0.171 ms |  30.65 ms |
| Updated  | 1               | 64       |  24.19 ms | 0.156 ms |  0.146 ms |  24.25 ms |
| Original | 1               | 512      |  30.48 ms | 0.150 ms |  0.140 ms |  30.56 ms |
| Updated  | 1               | 512      |  24.21 ms | 0.162 ms |  0.144 ms |  24.26 ms |
| Original | 32              | 64       | 225.99 ms | 8.697 ms | 25.643 ms | 241.60 ms |
| Updated  | 32              | 64       | 190.47 ms | 1.183 ms |  1.106 ms | 190.61 ms |
| Original | 32              | 512      | 224.84 ms | 4.470 ms | 13.040 ms | 227.92 ms |
| Updated  | 32              | 512      | 170.82 ms | 5.956 ms | 17.561 ms | 180.02 ms |
| Original | 512             | 64       | 231.52 ms | 4.592 ms |  5.104 ms | 233.16 ms |
| Updated  | 512             | 64       | 181.49 ms | 2.409 ms |  2.253 ms | 182.41 ms |
| Original | 512             | 512      | 230.62 ms | 4.596 ms |  9.071 ms | 234.37 ms |
| Updated  | 512             | 512      | 122.73 ms | 1.610 ms |  1.427 ms | 122.75 ms |

I doubt this will be revolutionary, but it's a simple change. Snapping to Microsoft.Extensions.ObjectPool would be better, but that's a more involved change.