Merged PR 795581: Fix theory detection in BXL XUnit test framework
Previously we detect if a test is a theory when there's a parameter.

If a test is serialize as `TestMe(value: 1)`, then our test framework thinks that it's a theory, and it will run it as
`RUN THEORY TestMe(value: 1)`. The problem with this is, if there are `[InlineData(1)]` and `[InlineData(2)]` for `TestMe`, we will have the following runs:
```
RUN THEORY TestMe(value: 1) =>
    RUN TEST TestMe(value: 1)(value: 1)
    RUN TEST TestMe(value: 1)(value: 2)
```

Similarly when the serialized test is `TestMe(value: 2)`.

Thus, instead of having 2 tests, we end up having 4 tests.

This PR also reenable QTest.

Related work items: #2195273