Merged PR 795555: Avoid caching real file system existence of paths that might be stale opaque outputs to be scrubbed by lazy materialization
We probe the real file system when doing a cache check and we can see an existing stale opaque output before the opaque directory is materialized from cache and said output is actually deleted. This can cause both cache misses and spurious DFAs.

This PR makes it so during cache lookup (and with lazy materialization enabled) we don't ever cache the existence of files that are under opaque directories.

Related work items: #2185421, #2185423