Merged PR 801793: Add checks to  flaky disabled HMC unit test to capture the state of LoadedCacheEntry task.
The unit test throws a NullRef exception on line 108. I am under the opinion that this could be because of this maybeLoadedCacheEntry1.Result.Value.MetadataHash. We are trying to access maybeLoadedCacheEntry1 whose cacheEntry could be null.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Engine/UnitTests/Scheduler/HistoricMetadataCacheTests.cs&version=GBmain&line=108&lineEnd=109&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents

In the TryGetCacheEntryAsync method of HistoricMetadataCache, we expect the cache entry to be present for the "unit test". As we have published that CacheEntry.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Engine/Scheduler/Caching/HistoricMetadataCache.cs&version=GBmain&line=236&lineEnd=237&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents

However, the method's definition suggests that if the entry is not found, it may return a null Result.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Engine/Cache/Fingerprints/TwoPhase/ITwoPhaseFingerprintStore.cs&version=GBmain&line=36&lineEnd=37&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents

Given this, I assume that due to some inconsistencies in the Linux environment, the fingerprint could be missing, which causes the code to skip the if condition and proceed to the TryGetCacheEntryAsync method in the base class. This results in a task being sucessful but has a null cache entry.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal?path=/Public/Src/Engine/Cache/Fingerprints/TwoPhase/InMemoryTwoPhaseFingerprintStore.cs&version=GBmain&line=83&lineEnd=84&lineStartColumn=1&lineEndColumn=1&lineStyle=plain&_a=contents

 Adding two additional asserts to confirm that the state of loadedCacheEntry is successful but the task has a null cache entry.

Related work items: #2191478