Merged PR 808205: Fix formatting of an npm auth token
Apparently, base64 encoding on linux automatically adds a newline every X characters. My speculation is that whatever tool reads that file does not expect this, so it reads a single line only, i.e., partial token value, so when it tries to auth with that partial value, it gets access denied.

Related work items: #2219577