Merged PR 810882: Handle retained engine output directories
Engine output directories are retained in some builds which is causing BXL to crash.

Ex: BuildXLCurrentLog directory which is to be deleted towards the end of the build is retained for the next build causing an issue.

This is handled by ensuring that we rename this retained folder and create a new directory.
Added a unit test to check the scenario.

Attaching the previous PR.
https://dev.azure.com/mseng/Domino/_git/BuildXL.Internal/pullrequest/805664

Related work items: #2188585