Merged PR 809409: Don't create Process class when dumping processes in job object
BuildXL is not able to dump a hung process due to "Process with an Id of N is not running" but the `GetExitCodeProcess` call just before didn't return success and it just created a handle so I'm pretty sure this case wasn't a race condition.  Using this PR's changes I confirmed that the process was in fact running and that `Process.GetProcessById` was too earnest in its diagnosis that the process was complete.  I was able to dump it without this method and so avoiding this call may be the best way to capture these kinds of process dumps.

Related work items: #2221778