Merged PR 815070: Fix version parsing of single digit versions in linux distro systems.
In the crash we find that the version is a single digit number which is causing the Version.ParseVersion method within TryParse to fail. It expects a "." to be present within the version. Hence converting the version from say 11 to 11.0 to create a valid version for TryParse to parse.

Related work items: #2215963