Merged PR 816459: EnsureLoadedAsync is synchronously waited on and must not queue up more async tasks
A deadlock can occur when the current context has all active threads running EnsureLoadedAsync which queues up an async task to that context.
To ensure this doesn't happen ensure we do not do any async/await in this method and ensure we return task result(s) that either are static or are from the singleton lazy initialization task.

This was subtly introduced with !797423

This should address #2229869 based on the dumps we have analyzed