Merged PR 814234: Make accommodations for shardsets in BXL
Make provisions for shardset domains. This is a side car change to the changes being made in the CB repo
https://mseng.visualstudio.com/Domino/_git/CloudBuild/pullrequest/815932

The idea behind this change is to accommodate for arbitrary values that a domain id maybe. The data type of domain id sort of leaked through into BXL and CB. That is implementation detail and `DomainIdFactory.Create` is responsible for managing that translation and should be used instead.

With this change, I expect CB and BXL to be able to pass in arbitrary string values to drop service.