Merged PR 816539: Use property instead of Task<bool> method to guard against new tasks being created inadvertently
This complements !816459. A deadlock was caused by the method `EnsureLoadedAsync()` queuing tasks (by virtue of awaiting async methods) into a dedicated task scheduler that was asynchronously waiting for `EnsureLoadedAsync` itself. That PR solved the issue by ensuring this method does not await and just returning `Task`s, and this PR makes it tighter by switching it to a property, which will never allow awaits