Merged PR 817081: Fix for TwoPhaseFailedToStoreMetadataForCacheEntry error bucket influx
Fix for TwoPhaseFailedToStoreMetadataForCacheEntry error bucket influx.
When the CancellationToken is requested, we log the TwoPhaseFailedToStoreMetadataForCacheEntry error and return a null entry.

But we do not want CancellationToken related issues to go under this bucket.  Also in the parent method we have a Contract which checks for the error to be logged if it receives a null entry. Handled this by adding an enum to include the three different states of the operation to store two-phase cache entry.

Related work items: #2231114