Merged PR 818475: Pass CB_COMPUTE_PROVIDER to DropDaemon
Specify that CB_COMPUTE_PROVIDER should get passed down to DropDaemon. This allows Windows builds to upload drops in CloudBuild on RM. Prior to this change they fail when trying to sign their build manifest because the child process that does that signing doesn't know which environment it is running in.