Merged PR 818162: Makes trace builder work when executed in VM
The current trace builder does not work when pips are executed in VM (for execution with admin privileges).

For executing pips in VM, BuildXL serializes the sandboxed process info, and sends it to the VM. In the VM, a dedicated sandbox executor is used to unwrap the sandboxed process info and execute it.

The path table is not included in the serialization of the sandboxed process info. Upon deserialization, the resulting sandboxed process info will have a brand-new path table.

When Detours reports a file access, the file access has a path, as a string, and a manifest path, as an `AbsolutePath` id. Note that this `id` is from the file access manifest constructed using the original path table before the sandboxed process info was sent to the VM. Thus, the id is only valid for the original path table, and not the one resulting from the deserialization.

The fix is to avoid using the manifest path. Instead, we reuse the "final" path re-created by the sandboxed process report. Note that sandboxed process report and the trace builder use the same path table. In this way, we no longer have to maintain the path block in the resulting trace; thus that block becomes deprecated.

Example build: https://cloudbuild.microsoft.com/build?id=5f5c0426-818f-4b22-9713-02e53ebb1150

The build still failed but it's no longer because of failure in saving the trace file.

Related work items: #2235757