Merged PR 819221: Remove impact of UnexpectedFileAccessesAreErrors from ExtraFingerprintSalts
When a process pip has an unexpected file access, its metadata is not stored to the cache and later builds cannot get cache hits on it. Therefore, it is not necessary for the unsafe UnexpectedFileAccessesAreErrors to additionally influence the fingerprint of the build.

Notable, this change does not require touching the existing tests in NonStandardOptionsTests that already validate builds cannot get cache hits on pips with unexpected file accesses. That shows that this change does not break that case.

Related work items: #2238691