Merged PR 819856: Revert "Merged PR 809634: Introduce EnableCpuAdaptiveSemaphore
Revert "Merged PR 809634: Introduce EnableCpuAdaptiveSemaphore

The CpuAdaptiveSemaphore feature enables the scheduler to throttle the number of processes to execute based on their historical CPU usage instead of slots. The UseHistoricalCpuUsage feature was already allowing us to use multiple slots per process; however, it has a few limitations:

1) We could not assign less than one slot per process, even if the process's CPU usage was less than 100% (one core).
2) We could not account for the BuildXL process's CPU usage. As more workers connected, we had to manually reduce concurrency because the increased distribution overhead caused higher CPU usage on the orchestrator. We were also using 0.9 * processorCount as the concurrency level because BuildXL itself consumes some CPU for hashing, cache lookups, etc.

With this new feature, we use the BuildXL process's CPU and RAM usages as semaphores to prevent oversaturating the machine during ongoing process executions. Every 2 seconds (the interval of the UpdateStatus method), the worker semaphores are updated with the current process and RAM usages. For remote workers, we pass the BuildXL process usage as part of the heartbeat messages sent every minute."

Reverted commit `2bb3dccb`.