Merged PR 819722: Pack nuget packages with nuget client SDK
Re-implements the pack command we use from nuget.exe into buildxl so that we can pack nuget packages without having to use msbuild or mono on Linux. The reason for not using msbuild is that it's hard to run under buildxl - especially multiple instances in parallel. This PR tries to keep our implementation as close to the nuget.exe implementation as possible so that we can have the same behaviour.

Related work items: #2237831