Merged PR 820777: Compress/Decompress execution log file
Adding changes to compress/decompress the  XLG file in order to reduce the file size and the upload time in ADO builds.

Using Zstd compression streams for compression and decompression of the streams.
A compression level of 2 has been tested and verified to be the optiomal level in this case.

Added decompression changes to ensure the xlg file can be used with the analyzers.
Added unit tests specific to Decompression changes

**Initial Testing**
Tested initially using standard .NET streams -DeflateStream,GZipStream, Brotli, Zlib
Schedule a total of 5 clean builds, one for each of the Compression streams and one base build.
Chose an Office build from OfficeGit_PR_LargeRegion.

- **Baseline Build**:
  - File size - 16.525 GB
  - No compression applied.
  - Total Build Time: 3:28:00.

- **GZip**:
  - File Size Reduction: 41.39% (6.84GB saved).
  - Build Time Increase: 5.32% (8m 23s added).
  - Total Build Time: 3:33:57.

- **Brotli**:
  - File Size Reduction: 76.76% (12.685GB saved).
  - Build Time Increase: 7.64% (12m 2s added).
  - Total Build Time: 3:38:10.

- **Deflate**:
  - File Size Reduction: 41.55% (6.88GB saved).
  - Build Time Increase: 11.58% (18m 40s added).
  - Total Build Time: 3:45:47.

- **Zlib**:
  - File Size Reduction: 41.60% (6.875GB saved).
  - Build Time Increase: 6.09% (9m 36s added).
  - Total Build Time: 3:36:59.

**Opted for Zstd Compression library.**
Tested it for 10 fully cached Office builds.

The results were impressive for compression level 2, which was considered optimal as the **file size reduced by 60%**.
The **original file size** was approximately **15.7GB**, and the **reduced file size** was **6.3GB**.

The **average duration of the product build** in minutes for **no compression** was **10.43**,
while for l**evel 2 compression**, it was **10.37**.

Hence the result of the testing was us opting for **ZSTD algorithm** with a **compression level of 2**.

Related work items: #2234228