Merged PR 821616: Add functionality to launch debugger from Detours code
Add functionality to launch debugger from Detours code.

When debugging Detours, we have so far relied on "printf". However, we would like to be able to see other data when debugging. For this, we need to be able to attach/launch a debugger when a detoured API is called.

Related work items: #2243051