Merged PR 823167: Fix scrubber path collapsing for Linux
We were doing a windows-specific handling of paths when collapsing them. That ended up with paths not being properly sorted, and paths to scrub that were not independent of each other. So in the end we tried to scrub the same directory twice and that produced a race

Related work items: #2244987