Merged PR 822403: [Detours]  Change detours so that it always faithfully returns the last error from the real API
Raw error code is the error code after calling the detoured API, while reported error code is the error code reported back to the sandbox client (the managed layer). One can think of the reported error code as the error code reported when running Procmon.

We distinguish these two kinds of error codes for the following reasons. A successful API call may not set the last error code to `ERROR_SUCCESS`. Detours does not want to modify error codes in order not to break users' apps, so it should inform the apps the value of error code right after the API call.

Consider the following example:

```cpp
DWORD attr = GetFileAttributesW(L"absent_file.txt");
BOOL createdDir = CreateDirectoryW(L"new_dir");
std::wcout << L"createdDir: " << createdDir << L", error code: " <<  GetLastErrorCode() << std::endl;
```

One may think that it will prints
```
createdDir: 1, error code: 0
```
because the directory creation succeeded. But in fact, it prints
```
createdDir: 1, error code: 2
```

Unfortunately, the sandbox client relies on the fact that if an API call is successful, then its error code is expected to be `ERROR_SUCCESS`. Thus, for the reported error code, Detours ensures that it is `ERROR_SUCCESS` when the API call returns `TRUE`.

For accountability, both error codes are reported back to the sandbox client, but the client should only rely on the reported error code.

This PR also made some changes wrt. denied access behavior. Before, we report access before we decide whether the access should be denied or not. So, it can happen that the client sees that the error code is ERROR_SUCCESS, but the API exits with DENIED_ERROR. The change is to delay reporting file access until after the decision to deny the access or not is made, and this change updates a number of functions to conform to this as a standard

Example validation :
https://dev.azure.com/mseng/Domino/_build/results?buildId=29547961&view=results

Related work items: #2234559