Merged PR 823947: Compress xlg files with zstd
This reapplies an earlier change to compress execution log files with zstd as they are written. The more involved changes are in the execution analyzer which must decompress the execution log as it is read. A decompressed file is written to the temp directory and cache there for subsequent analyzer runs against the same input xlg file. Files are not decompressed in memory on the fly because the decompressed stream does not offer seeking, which the execution analyzer framework currently requires.

Original PR: !820777