Merged PR 824233: Avoid scrubbing sources when a hardlink points to it
When a tool produces a hardlink to a source file as a shared opaque output, we flag the output but modify the source file as well.  On the next run, if the scrubber finds the source file it will delete it as well.
This is a temporary solution to patch this problem. If an output has more hardlinks than expected we just avoid flagging it. This avoids the source deletion (but leaves the hardlink as a stale output for the next build). A proper fix is tracked [here](https://dev.azure.com/mseng/1ES/_workitems/edit/2167806), this is just a quick solution to avoid deleting source files.

Related work items: #2244989